requirement.txt
langchain
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
